feat: add redis queue

+++ DotNetApi/DTOs/RoutesDriverDto.cs
@@ -2,6 +2,6 @@ namespace DotNetApi.DTOs;
 public class RoutesDriverDto
 {
     public string RouteId { get; set; }
-    public int Lat { get; set; }
-    public int Lng { get; set; }
+    public float Lat { get; set; }
+    public float Lng { get; set; }
 }
